requirements.txt
ipykernel==4.6.1
ipython==6.1.0
ipywidgets==7.0.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
matplotlib==2.0.2
mistune==0.7.4
nbconvert==5.3.1
notebook==5.0.0
numpy==1.13.1
scipy==0.19.1
tornado==4.5.2
google-auth==1.0.2
google-cloud-core==0.27.1
google-cloud-storage==1.4.0
google-resumable-media==0.2.3
googleapis-common-protos==1.5.2
